% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEST5EXP.R
\name{TEST5EXP}
\alias{TEST5EXP}
\title{Gives test statistic value and critical value for one parameter exponential distribution with group size =5}
\usage{
TEST5EXP(x1, x2, x3, x4, x5, alpha)
}
\arguments{
\item{x1}{is sample from Exp Population 1.}

\item{x2}{is sample from Exp Population 2.}

\item{x3}{is sample from Exp Population 3.}

\item{x4}{is sample from Exp Population 4.}

\item{x5}{is sample from Exp Population 5.}

\item{alpha}{is numeric value (level of significance)}
}
\description{
Gives test statistic value and critical value for one parameter exponential distribution with group size =5
}
